+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.14.0] - 2025-01-20
 
 ### Breaking
 
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added derivable trait [`ErrorWithStatusCode`](https://docs.rs/picoserve/latest/picoserve/response/trait.ErrorWithStatusCode.html) to facilitate creating error responses. Deriving `ErrorWithStatusCode` also derives [`IntoResponse`](https://docs.rs/picoserve/latest/picoserve/response/trait.IntoResponse.html)
+- Added derivable trait [`ErrorWithStatusCode`](https://docs.rs/picoserve/0.14.0/picoserve/response/trait.ErrorWithStatusCode.html) to facilitate creating error responses. Deriving `ErrorWithStatusCode` also derives [`IntoResponse`](https://docs.rs/picoserve/0.14.0/picoserve/response/trait.IntoResponse.html)
 
 ## [0.13.3] - 2024-12-26